backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
openai==1.3.0
jinja2==3.1.2
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
authlib==1.3.0
itsdangerous==2.1.2
sentry-sdk[fastapi]==1.39.1
httpx==0.25.2